Update pkgutil to use importlib.import_module.
If `pkgutil.get_data` fails to find the given package in `sys.modules` it will attempt to load it with `importlib._bootstrap._load(spec))`.  However, this seems to cause issues when a child of the given module is imported normally later (as described in this [Stack Overflow question](https://stackoverflow.com/questions/59377661/python-pkgutil-get-data-disrupts-future-imports).

I'm not too familiar with how the overall system works, but replacing the line with `importlib.import_module(package)` seems to work.